+++ lumispy/signals/luminescence_spectrum.py
@@ -28,6 +28,7 @@ from hyperspy.axes import DataAxis
 from lumispy.signals.common_luminescence import CommonLumi
 from lumispy import nm2invcm, to_array, savetxt
 from lumispy.utils.axes import GRATING_EQUATION_DOCSTRING_PARAMETERS
+from lumispy.utils.signals import com
 from lumispy.utils import (
     axis2eV,
     data2eV,
@@ -117,12 +118,15 @@ class LumiSpectrum(Signal1D, CommonLumi):
         # create new signal object with correct data, axes, metadata
         else:
             s2 = self._deepcopy_with_new_data(
-                s2data, copy_variance=(not jacobian), copy_learning_results=False
+                s2data,
+                copy_variance=(not jacobian),
+                copy_learning_results=False,
             )
         # convert axis
         oldaxis = self.axes_manager.signal_axes[0]
         s2.axes_manager.set_axis(
-            newaxis, self.axes_manager.signal_axes[0].index_in_axes_manager
+            newaxis,
+            self.axes_manager.signal_axes[0].index_in_axes_manager,
         )
         # convert variance
         if self.metadata.has_item("Signal.Noise_properties.variance"):
@@ -314,7 +318,8 @@ class LumiSpectrum(Signal1D, CommonLumi):
             s2 = self.to_invcm(inplace=inplace, jacobian=jacobian)
         # replace axis
         s2.axes_manager.set_axis(
-            invcmaxis, self.axes_manager.signal_axes[0].index_in_axes_manager
+            invcmaxis,
+            self.axes_manager.signal_axes[0].index_in_axes_manager,
         )
         s2.data = s2.isig[::-1].data
         # replace variance axis
@@ -322,7 +327,8 @@ class LumiSpectrum(Signal1D, CommonLumi):
             s2.get_noise_variance(), (float, int)
         ):
             s2.metadata.Signal.Noise_properties.variance.axes_manager.set_axis(
-                invcmaxis, s2.axes_manager.signal_axes[0].index_in_axes_manager
+                invcmaxis,
+                s2.axes_manager.signal_axes[0].index_in_axes_manager,
             )
             s2.metadata.Signal.Noise_properties.variance.data = (
                 s2.metadata.Signal.Noise_properties.variance.isig[::-1].data
@@ -330,7 +336,10 @@ class LumiSpectrum(Signal1D, CommonLumi):
         if not inplace:
             return s2
 
-    to_invcm_relative.__doc__ %= (TO_INVCM_DOCSTRING, TO_INVCMREL_EXAMPLE)
+    to_invcm_relative.__doc__ %= (
+        TO_INVCM_DOCSTRING,
+        TO_INVCMREL_EXAMPLE,
+    )
 
     # Alias Method Name
     to_raman_shift = to_invcm_relative
@@ -451,7 +460,13 @@ class LumiSpectrum(Signal1D, CommonLumi):
     """
 
     def savetxt(
-        self, filename, fmt="%.5f", delimiter="\t", axes=True, transpose=False, **kwargs
+        self,
+        filename,
+        fmt="%.5f",
+        delimiter="\t",
+        axes=True,
+        transpose=False,
+        **kwargs,
     ):
         """Writes luminescence spectrum object to simple text file.
         %s
@@ -460,7 +475,11 @@ class LumiSpectrum(Signal1D, CommonLumi):
         """
         savetxt(self, filename, fmt, delimiter, axes, transpose, **kwargs)
 
-    savetxt.__doc__ %= (SAVETXT_DOCSTRING, SAVETXT_PARAMETERS, SAVETXT_EXAMPLE)
+    savetxt.__doc__ %= (
+        SAVETXT_DOCSTRING,
+        SAVETXT_PARAMETERS,
+        SAVETXT_EXAMPLE,
+    )
 
     TOARRAY_EXAMPLE = """
     Notes
@@ -502,7 +521,11 @@ class LumiSpectrum(Signal1D, CommonLumi):
         """
         return to_array(self, axes, transpose)
 
-    to_array.__doc__ %= (TOARRAY_DOCSTRING, TOARRAY_PARAMETERS, TOARRAY_EXAMPLE)
+    to_array.__doc__ %= (
+        TOARRAY_DOCSTRING,
+        TOARRAY_PARAMETERS,
+        TOARRAY_EXAMPLE,
+    )
 
     def px_to_nm_grating_solver(
         self,
@@ -553,7 +576,8 @@ class LumiSpectrum(Signal1D, CommonLumi):
             s = self.deepcopy()
 
         s.axes_manager.set_axis(
-            nm_axis, self.axes_manager.signal_axes[0].index_in_axes_manager
+            nm_axis,
+            self.axes_manager.signal_axes[0].index_in_axes_manager,
         )
 
         return s
@@ -561,6 +585,62 @@ class LumiSpectrum(Signal1D, CommonLumi):
     px_to_nm_grating_solver.__doc__ %= GRATING_EQUATION_DOCSTRING_PARAMETERS.replace(
         "\n", "\n\t"
     )
+
+    def centroid(self, signal_range=None, **kwargs):
+        """
+        Finds the centroid (center of mass) of a peak in the spectrum from
+        the wavelength (or pixel number) and the intensity at each pixel
+        value. It basically represents a "weighted average" of the peak.
+
+        Notes
+        -----
+        This function only works for a single peak. If you have multiple
+        peaks, slice the signal beforehand or use the signal_range parameter.
+
+        TODO: Implement this function for multiple peaks (with the npeaks
+        parameter) by finding the top 2 peaks from mean spectrum and then
+        returning a signal with 2 com.
+
+        Parameters
+        ----------
+        signal_range : tuple of ints or floats, optional
+            A tuple representing the indices of the signal axis (start index,
+            end index) where the peak is located. If the tuple contains int,
+            it slices on index. If the tuple contains float, it slices on
+            signal units (default HyperSpy s.inav[:] functionality).
+
+        kwargs : dictionary
+            For the scipy.interpolate.interp1d function.
+
+        Returns
+        -------
+        signal : BaseSignal
+            A BaseSignal with signal dimension of 0, where at each navigation
+            index there is a scalar value containing the center of mass for
+            each navigation pixel.
+        """
+        if signal_range:
+            if type(signal_range) != tuple:
+                raise TypeError(
+                    "The `signal_range` parameter must be a tuple of length 2."
+                )
+            if len(signal_range) != 2:
+                raise ValueError(
+                    f"The `signal_range` parameter must be a tuple of length 2. "
+                    "You passed a tuple of length {len(signal_range)}."
+                )
+
+            s = self.isig[signal_range[0] : signal_range[1]]
+
+        else:
+            s = self
+
+        signal_axis = s.axes_manager.signal_axes[0]
+        center_of_mass = s.map(com, signal_axis=signal_axis, inplace=False)
+
+        # Transfer axes metadata to title
+        center_of_mass.metadata.General.title = f"Centroid map of {signal_axis.name} ({signal_axis.units}) for {center_of_mass.metadata.General.title}"
+        return center_of_mass
 
 
 class LazyLumiSpectrum(LazySignal, LumiSpectrum):